fix: makes sorting the edges work with a hack
The base sorting mechanism of Hornet is completely broken, I have provided a simpler implementation
of the segmented sort that uses only cub and I have initialized the edge data "vids" array to a
maximum value to prevent zeros from being put in front of the adjlist.

A real solution will need to apply segmentation sort to non continous regions of the array, in order
to skip all trailing zeros in the adjlist's block.